Split config into browser-compatible and Node.js modules
Fix Firefox extension build by separating browser-compatible config
(types, defaults, FIELDS) from Node.js-specific code (ConfigManager,
env parsing, CLI generation).

- configDefaults.ts: Browser-compatible types, FIELDS, DEFAULTS
- config.ts: Node.js code that re-exports from configDefaults
- core.ts: Import from configDefaults for browser compatibility
- index.ts: Export PROVIDERS, BROWSERS, etc. for consumer use
- Add clean script to package.json
- Update CI to clean before building
- Update extension prebuild to clean before building
