+++ extension/package.json
@@ -12,9 +12,9 @@
     "dev:chrome:tmp-profile": "wxt -b chrome",
     "dev:firefox": "node scripts/mkdirp.js .wxt/firefox-data && WEB_EXT_FIREFOX_PROFILE=.wxt/firefox-data WEB_EXT_KEEP_PROFILE_CHANGES=true wxt -b firefox",
     "dev:firefox:tmp-profile": "wxt -b firefox",
-    "prebuild:chrome": "cd .. && pnpm build",
+    "prebuild:chrome": "cd .. && pnpm clean && pnpm build",
     "build:chrome": "wxt build -b chrome",
-    "prebuild:firefox": "cd .. && pnpm build",
+    "prebuild:firefox": "cd .. && pnpm clean && pnpm build",
     "build:firefox": "wxt build -b firefox",
     "zip:chrome": "wxt zip -b chrome",
     "zip:firefox": "wxt zip -b firefox",
